feat: implement scope utility functions for validation and extraction
